+++ DesignerRHN/DesignerRHN-ejb/src/java/Entidades/Paises.java
@@ -31,7 +31,7 @@ import javax.xml.bind.annotation.XmlTransient;
     @NamedQuery(name = "Paises.findAll", query = "SELECT p FROM Paises p")})
 public class Paises implements Serializable {
     @Column(name = "CODIGO")
-    private Short codigo;
+    private Integer codigo;
     // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
     private static final long serialVersionUID = 1L;
     @Id
@@ -109,11 +109,11 @@ public class Paises implements Serializable {
         return "Entidades.Paises[ secuencia=" + secuencia + " ]";
     }
 
-    public Short getCodigo() {
+    public Integer getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(Short codigo) {
+    public void setCodigo(Integer codigo) {
         this.codigo = codigo;
     }
